working POF - data path subscribing instrument
A numerical instrument DashboardInstrument_Single is derived to a new EngineI class, which is working the other way around for data than its base class: it gets in its constructor a list of signal data path it can subscribe to. It gets also language dependent description of those. The class is demonstrating how a generic instrument classes like this can be created so  they are more generic possible, and then by some user selection (not implemented) they customize themselves to the data the user wants to monitor. In a similar manner, a dial can be made, graphs, etc.

This is totally working example: by creating as many Engine RPM instruments as you like in a floating instrument panes, they all subscribe (randomly) to one of the available three signal paths (oil pressure, engine coolant temperature, engine speed).

This POC demonstrates that indivual floating instruments can be made so that they subscribe to a multiplexer, which is running either on the OpenCPN API or in the plug-in.   In a similar manner, the SIgnal K parser can work either under OpenCPN or in the plug-in -the instrument is not aware of this, of course.

This principle does remove all the need for the 32 or even 64 bit limitation dictated by "instrument capacity flag" which is the limiting factor of the current design.

What does not work is the destructor removing the call-back: this works but somehow the asynchronous data event arrives in between (there are semaphores?) and a call to a call-back of a disappeared object is made and the program crashes. To be investigated (in the end of tactics_pi.cpp).

C++ 2011 containers have been used extensively, They are not the fastest ones. Boost-library could be used, which would allow even to define lambdas straight from the constructor. There are also other templates around in the interweb.