Back to TacticsWindow (pane) hosted call-back STL container
This is more satisfying implementation, taking into account the nature of windows rearranging (addig or removing instruments), the DashboardWindow object (pane), deriving from TacticsWindow class now first EVT_CLOSEs all the instrument windows (allowing them to unsubscribe), the destroys them (by delete). After which, the dashboard_pi will close and destroy the window pane object itself. The STL container is now gone. It will be created again when the window pane is created again and populated with the new content. This way, it remains small and fast-to-search, instead of having all the instruments call-back in the same bucket.