Successful end of dev200CBack FF. Read->
Let's halt this flying fish branch - without pulling it down to dev200, perhaps better to create another FF: this branch has proven the call-back method working perfectly, resolved cross-compliation problems, defined that we need to use the latest OS if we want to use WebView and I think we shoudl: the instrujs.cpp actually replaces perfectly enginei.cpp demo class, it can also show numerical values, no problem - this is the demo for that! Remains many ideas to develop how to configure and how to make persistant that configuration. Also, one needs to find a scalable dial JS. But that is the matter of another branch, this branch says nows 'bye'...