Event handler for docking/undocking
Now can detect and deal with the docking / undocking events, using the (only) available AUI manager render event and by having own book keeping. This allows to rebuild the freshly docked pane, which is needed if there is a WebView window. No need to rebuild when undocking the pane, though, but the event detection is there and just registered in ApplyConfig(), in case. Remains the issue with the sized docked pane restart with WebView based instrument, it gets the default size no matter what is the width of the vertically docked pane on the right.