Separate MIN/MAX macros for H and V
Careful adjustement of the sizes is needed, one
needs to compare the enginedjg.css with the proportions
defined in engineedjg.h - the heights have the same
elements and the height proportions are good to keep the
same, without going to decimals. For the width, one needs
to take into account the horizontal arranged instruments:
In Dashboard one cannot resize the horizontal size, so the
switch needs to be adjusted a bit large to allow the path
to displayed (on the Linux WebKit 3.0/wxWidgets 3.0, other
platforms can take advantage of the CSS resizable fonts
depending of the viewport size.