Stopping this thread, success (read report ->)
Webpack is not for half hearted! Coming from C++ it _not_ CMake but rather CMake _and_ linker, together. The compiler is in the browser... Tough to make a mental switch! To help in that, I stopped adding semicolons (;) since the standard interpreter adds them automatically. This helps to make the mental switch, it does not make your code any slower or faster or smaller or bigger! You'll find a lot of heated debate about it, but my choice is to perhaps to move to TypeScript later on and have kind of a pre-compilation not only Webpack packing.

The success of this thread is that now I know how to interface with the public window area and thus with C++ WebView script execution (iface.js) - need to move that to ../../src . Also from that public class I can send events to the modules which can come and pick the data up from the object.

Using events: moving to use some state machine framework. Tried and detected a compatibility issue with IE (sigh!) with xstate framworks. That's really a pity, but need to move backward for the IE now and try some older frameworks.