Resolves slow start crash when data comes in
When AUI is (slowly rearranging the instrument windows at startup, they get recreated. If, at this moment data is coming in from fast channel like socket, data is sent out to a window (instrument) which was just removed to be recreated. Crash! Do not send data to anybody at start while AUI  window rearranging is going on!