feat: allow and base all time on GNSS (GPS) time
Explanation: When reeling back some old recordings
which have GNSS (GPS) time data, the local (CPU) time
and the timestamps are completely different.
The derived data (like Tactics) need to get its time-
stamps but using local time is now not possible,
the difference is too great and it would be impossible
to find data anymore.
The solution is to use GNSS (GPS) data everywhere,
_if_ it is available. There are some issues with that:
the messages making that "clock" to tick are too
slow. Therefore we still need to use the CPU clock
in between the two GNSS (GPS) sentences.
This commit is doing all the above and has been
tested with aavan2k.data and the usual NMEA-0183
recordings.