fix: #154 (docked add/remove crash win10)
The InstruJS class of instruments needs the Close()
event and Destroy() (destructor) in ordered manner
so that GTK3 created threads for wxWebView are
properly killed on Linux before the window.
Helas, on Windows if one Destroy() a docked
pane for its recreation and the Update() the AUI
this crash occurs. As in the ticket #154, the
workaround is to undock (detach) the pane first.
This patch is doing that on behalf of the user.
One may note an extra AUI Update() before
the Dashboard window Destroy().
That seems to do an adequate fix for now.